requirements.txt
asgiref==3.8.1
Django==5.2.1
django-redis==5.4.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
Faker==37.1.0
PyJWT==2.9.0
python-decouple==3.8
redis==6.0.0
sqlparse==0.5.3
tzdata==2025.2
